flow: make JS metatype a module
Move the JS metatype code outside src/lib/flow/ into
src/modules/flow-metatype/js/, allowing it to be compiled as a dynamic
library (external module).

The code was moved from the sol-flow-js.c implementation file. No major
changes except for minor tweaks so that the header file is removed. The
boilerplate for the module was added at the end to avoid further
unrelated changes.

Since the API specific to JS metatype is not exposed anymore, the
test-javascript which is written in C, was changed to use
sol_flow_parse_string_metatype() instead.

The config option JAVASCRIPT was replaced with FLOW_METATYPE_JAVASCRIPT
and moved to the appropriate section in the menu.

The Makefile in src/thirdparty which triggered the compilation of
duktape was removed. Now duktape is compiled as part of the JS module.
