+++ src/lib/flow/Makefile
@@ -12,9 +12,6 @@ obj-flow-$(NODE_DESCRIPTION) += \
     sol-flow-resolver.o \
     sol-flow-builder.o
 
-obj-flow-$(JAVASCRIPT) += \
-	sol-flow-js.o
-
 obj-flow-$(ENABLE_DYNAMIC_MODULES) += \
     sol-flow-parser-dynamic.o
 
@@ -33,7 +30,6 @@ headers-$(FLOW_SUPPORT) := \
     include/sol-flow-builder.h \
     include/sol-flow.h \
     include/sol-flow-inspector.h \
-    include/sol-flow-js.h \
     include/sol-flow-metatype.h \
     include/sol-flow-packet.h \
     include/sol-flow-parser.h \
