+++ src/modules/flow-metatype/js/Kconfig
@@ -1,0 +1,6 @@
+config FLOW_METATYPE_JAVASCRIPT
+	tristate "JavaScript flow metatype"
+	depends on FLOW_SUPPORT && PLATFORM_LINUX && HAVE_DUKTAPE_SRC
+	default m if MODULES=y
+	help
+	    Allow creating node types using JavaScript
